Update Main.py
Added a basic menu system and different game states; allows the game to reset after the game is over.  Moved rendering of ball and paddle to their respective classes.  Moved the initialization of ball and paddle variables to separate functions.  Added code to test if the menu buttons are clicked which updates the game state respectively.